// src/app/overview/enums/popup.ts
export enum PopupType {
  HaveCard = "haveCard",
  Rates = "rateAges",
  Educate = "educate",
  Type = "types",
  UseApp = "useApp",
  Benefit = "benefit",
  Career = "career",
  IncomeNum = "incomeNum",
  Income = "income",
  Career2 = "career2",
  Skill1 = "skill1",
  Skill2 = "skill2",
  Skill3 = "skill3",
  Skill4 = "skill4",
  Health1 = "health1",
  Health2 = "health2",
  Health3 = "health3",
  Health4 = "health4",
  Health5 = "health5",
  Home1 = "home1",
  Home2 = "home2",
}